refactor(cmd): merge coldstep-action into the combined coldstep binary
Ship one artifact instead of two. cmd/coldstep-action becomes internal/actioncli
(package main -> package actioncli; func main -> Dispatch returning an exit
code). The coldstep binary's CLI dispatcher routes run/validate as before and
start/stop/diff/assert-integrity into actioncli.Dispatch. Consumers download a
single prebuilt coldstep binary; the composite invokes 'coldstep stop' etc.
instead of a separate bin/coldstep-action.

Build/CI/supply-chain/TS/docs repoints follow in subsequent commits.
