+++ cmd/coldstep/main.go
@@ -4,6 +4,7 @@ import (
 	"fmt"
 	"os"
 
+	"github.com/coldstep-io/coldstep/internal/actioncli"
 	"github.com/coldstep-io/coldstep/internal/agent"
 )
 
@@ -16,7 +17,7 @@ func main() {
 
 func runCLI(args []string) int {
 	if len(args) < 2 {
-		fmt.Fprintln(os.Stderr, "usage: coldstep <run|validate> [args...]")
+		fmt.Fprintln(os.Stderr, "usage: coldstep <run|validate|start|stop|diff|assert-integrity> [args...]")
 		return 2
 	}
 	switch args[1] {
@@ -29,6 +30,11 @@ func runCLI(args []string) int {
 	case "validate":
 		return runValidate(args[2:], os.Stdout, os.Stderr)
 	default:
+		// start / stop / diff / assert-integrity — the composite action helper,
+		// merged into this binary so the Release ships a single artifact.
+		if _, ok := actioncli.Commands[args[1]]; ok {
+			return actioncli.Dispatch(args[1:])
+		}
 		fmt.Fprintln(os.Stderr, "unknown command")
 		return 2
 	}
