+++ internal/actioncli/agent_pid_linux.go
@@ -1,6 +1,6 @@
 //go:build linux
 
-package main
+package actioncli
 
 import (
 	"fmt"
