+++ internal/actioncli/allowlist_files.go
@@ -1,4 +1,4 @@
-package main
+package actioncli
 
 import (
 	"fmt"
