+++ internal/actioncli/main_linux_test.go
@@ -1,6 +1,6 @@
 //go:build linux
 
-package main
+package actioncli
 
 import (
 	"os"
