+++ internal/actioncli/main_test.go
@@ -1,4 +1,4 @@
-package main
+package actioncli
 
 import (
 	"bytes"
