+++ internal/actioncli/report_markdown.go
@@ -1,4 +1,4 @@
-package main
+package actioncli
 
 import (
 	"flag"
